added restart alert when changing language
